Add UDP extension for Tap custom sink
