+++ CODEOWNERS
@@ -259,9 +259,10 @@ extensions/filters/http/oauth2 @derekargueta @mattklein123
 /*/extensions/filters/http/rbac @yangminzhu @yanavlasov
 /*/extensions/filters/common/rbac @yangminzhu @yanavlasov
 # tap
-/*/extensions/filters/http/tap @mattklein123 @xu1zhou
-/*/extensions/common/tap @mattklein123 @xu1zhou
-/*/extensions/transport_sockets/tap @mattklein123 @xu1zhou
+/*/extensions/filters/http/tap @mattklein123 @coolg92003 @xu1zhou
+/*/extensions/common/tap @mattklein123 @coolg92003 @xu1zhou
+/*/extensions/transport_sockets/tap @mattklein123 @coolg92003 @xu1zhou
+/*/extensions/tap_sinks/udp_sink @wbpcode @coolg92003 @yiyibaoguo
 # local rate limit
 /*/extensions/filters/network/local_ratelimit @mattklein123 @wbpcode
 /*/extensions/filters/listener/local_ratelimit @mattklein123 @JuniorHsu
