+++ api/envoy/extensions/tap_sinks/udp_sink/v3/udp_sink.proto
@@ -1,0 +1,22 @@
+syntax = "proto3";
+
+package envoy.extensions.tap_sinks.udp_sink.v3;
+
+import "envoy/config/core/v3/address.proto";
+
+import "udpa/annotations/status.proto";
+
+option java_package = "io.envoyproxy.envoy.extensions.tap_sinks.udp_sink.v3";
+option java_outer_classname = "UdpSinkProto";
+option java_multiple_files = true;
+option go_package = "github.com/envoyproxy/go-control-plane/envoy/extensions/tap_sinks/udp_sink/v3;udp_sinkv3";
+option (udpa.annotations.file_status).package_version_status = ACTIVE;
+
+// [#protodoc-title: Udp sink configuration]
+// [#extension: envoy.tap_sinks.udp_sink]
+
+// Udp sink configuration.
+message UdpSink {
+  //   Configure UDP Address
+  config.core.v3.SocketAddress udp_address = 1;
+}
