+++ changelogs/current.yaml
@@ -21,6 +21,9 @@ new_features:
   change: |
     Added :ref:`virtualClusterName() <config_http_filters_lua_stream_info_virtual_cluster_name>` API to the Stream Info
     Object to get the name of the virtual cluster matched.
+- area: tap
+  change: |
+    Added an UDP extension for tap custom sink.
 - area: udp_proxy
   change: |
     Added support for outlier detection in UDP proxy. This change can be temporarily reverted by setting runtime guard
