+++ docs/root/api-v3/config/tap_sinks/tap_sinks.rst
@@ -1,0 +1,8 @@
+Transport sockets
+=================
+
+.. toctree::
+  :glob:
+  :maxdepth: 2
+
+  ../../extensions/tap_sinks/*/v3/*
